+++ fuzz/fuzz_targets/object_header_parse.rs
@@ -1,9 +1,7 @@
 #![no_main]
 
-// Scaffold target for future cargo-fuzz/libFuzzer admission.
-//
-// Intended harness body:
-//
-// fuzz_target!(|data: &[u8]| {
-//     let _ = sagnir_object::parse_object_header(data);
-// });
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = sagnir_object::parse_object_header(data);
+});